SIMPLEX-153 Refactor and fix tests.
